# Update WASM cache to store Result instead