--- a/core/wasm/src/types.rs
+++ b/core/wasm/src/types.rs
@@ -1,7 +1,7 @@
 use primitives::types::{PromiseId, AccountId, Balance, Mana, BlockIndex};
 use wasmer_runtime::error as WasmerError;
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 /// Error that can occur while preparing or executing wasm smart-contract.
 pub enum PrepareError {
     /// Error happened while serializing the module.
@@ -146,7 +146,7 @@ impl ::std::fmt::Display for RuntimeError {
 }
 
 /// Wrapped error
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub enum Error {
     /// Method name can't be decoded to UTF8.
     BadUtf8,
@@ -163,7 +163,7 @@ pub enum Error {
 
     Prepare(PrepareError),
 
-    Cache(WasmerError::CacheError),
+    Cache(String),
 }
 
 impl From<WasmerError::Error> for Error {