Fixes #568 
Also adding a benchmark to try run some bad code.

Before this commit:
```
test runtime_send_money                 ... bench:     472,202 ns/iter (+/- 27,107)
test runtime_wasm_bad_code              ... bench:   4,042,358 ns/iter (+/- 269,702)
test runtime_wasm_benchmark_10_reads    ... bench:   1,441,188 ns/iter (+/- 168,624)
test runtime_wasm_benchmark_sum_1000    ... bench:   1,257,166 ns/iter (+/- 118,543)
test runtime_wasm_benchmark_sum_1000000 ... bench:  14,212,002 ns/iter (+/- 328,197)
test runtime_wasm_set_value             ... bench:   1,286,205 ns/iter (+/- 52,033)
```

After this commit:
```
test runtime_send_money                 ... bench:     473,913 ns/iter (+/- 42,013)
test runtime_wasm_bad_code              ... bench:     804,576 ns/iter (+/- 46,408)
test runtime_wasm_benchmark_10_reads    ... bench:   1,424,945 ns/iter (+/- 133,040)
test runtime_wasm_benchmark_sum_1000    ... bench:   1,306,655 ns/iter (+/- 188,243)
test runtime_wasm_benchmark_sum_1000000 ... bench:  14,121,906 ns/iter (+/- 1,260,672)
test runtime_wasm_set_value             ... bench:   1,274,420 ns/iter (+/- 66,046)
```