@evgenykuzyakov 2019-02-14
Fixes #568